fix(ir): verify For/DoUntil bodies; make Capture.index the wire's u32 (bug-33)
- `verify_package`'s structural walk recursed into `If`/`While`/`ForEach`/`Trap`/
  `Match` bodies but not `For`/`DoUntil`, so the empty-`MATCH` invariant was never
  applied inside those loops. (The semantic pass does recurse, so this was a
  defense-in-depth gap, not a soundness hole.)
- `IrValue::Capture.index` was a `usize` encoded with `*index as u32` and decoded
  back to `usize`: an index above `u32::MAX` silently round-tripped to a different
  slot. Narrow the field to `u32` — the width the format actually stores — so
  encode and decode cannot disagree. NIR keeps `usize` for offset arithmetic.
- Tests: an empty MATCH inside a `For` and inside a `DoUntil` is now rejected by
  the structural pass, and `u32::MAX` round-trips exactly.

+++ src/ir/value.rs
@@ -43,7 +43,10 @@ pub(crate) enum IrValue {
         captures: Vec<IrValue>,
     },
     Capture {
-        index: usize,
+        /// The closure environment slot this capture reads. `u32` is the width
+        /// the package format encodes, so the in-memory value cannot silently
+        /// disagree with its serialization.
+        index: u32,
         type_: String,
         /// When set, the env slot at `index` holds a pointer to the parent
         /// binding's slot (a non-escaping `MUT` borrow), so the capture binds a
